// frontend/src/app/core/generated/model/claim.ts
/**
 * Claims Management System
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ClaimStatus } from './claim-status';
import { Damage } from './damage';


export interface Claim { 
    id: string;
    title: string;
    description?: string;
    status: ClaimStatus;
    /**
     * Calculated by the backend
     */
    totalAmount: number;
    damages: Array<Damage>;
}
export namespace Claim {
}