// frontend/src/app/core/generated/model/damage.ts
/**
 * Claims Management System
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DamageSeverity } from './damage-severity';


export interface Damage { 
    id: string;
    part: string;
    severity: DamageSeverity;
    imageUrl: string;
    price: number;
}
export namespace Damage {
}